ML/requirements.txt
uvicorn
fastapi
numpy
scikit-learn
seaborn
mysql-connector-python